+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 
+* Added support for secure websockets with CPython 2.7.0 - 2.7.8
 * Fixed RoleSecretAuthDelegate name in docstrings about authentication
 
 v1.0.0 (2017-03-07)